+++ zkvm/src/lib.rs
@@ -20,7 +20,8 @@ pub use self::errors::VMError;
 pub use self::ops::{Instruction, Opcode};
 pub use self::predicate::{Predicate, PredicateWitness};
 pub use self::prover::Prover;
+pub use self::signature::VerificationKey;
 pub use self::txlog::{Entry, TxID, UTXO};
-pub use self::types::{Commitment, Data, Item, Value, WideValue};
+pub use self::types::{Commitment, Data, DataWitness, Item, Value, WideValue};
 pub use self::verifier::Verifier;
 pub use self::vm::{Tx, VerifiedTx};
