+++ zkvm/src/vm.rs
@@ -693,40 +693,3 @@ where
         .map_err(|_| VMError::R1CSInconsistency)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn issue() {
-        let privkey = Scalar::random(&mut rand::thread_rng());
-        let pubkey = VerificationKey::from_secret(&privkey);
-
-        let predicate = match Predicate::from_witness(PredicateWitness::Key(privkey)) {
-            Err(_) => return assert!(false),
-            Ok(x) => x,
-        };
-
-        // pushdata qty
-        // var
-        // pushdata flv
-        // var
-        // pushdata pred <- pubkey to issue to
-        // issue
-        let mut program = vec![(
-            // TBD: replace with real scalars
-            Instruction::Push(Data::Witness(DataWitness::Scalar(Box::new(Scalar::one())))),
-            Instruction::Var,
-            // TBD: replace with real scalars
-            Instruction::Push(Data::Witness(DataWitness::Scalar(Box::new(Scalar::one())))),
-            Instruction::Var,
-            // Predicate
-            Instruction::Push(Data::Witness(DataWitness::Predicate(Box::new(predicate)))),
-            Instruction::Issue,
-            Instruction::Signtx,
-        )];
-
-
-    }
-}
